backend/requirements.txt
# ═══════════════════════════════════════════════════════════════════════════
# DIGITAL DENIS — Backend Requirements
# ═══════════════════════════════════════════════════════════════════════════

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.4.2
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# LLM
httpx==0.26.0
openai==1.10.0  # For OpenRouter compatibility
groq==0.4.0   # For Voice Transcription

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Background Jobs
celery==5.3.6

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
PyYAML==6.0.1
python-slugify==8.0.4

# Analytics & ML
scikit-learn==1.4.0
hdbscan==0.8.33


# Observability
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pywebpush==1.14.3
